Feature: gradually fall back to other servers if a server is yielding errors
This commit contains a few entangled changes. Firstly, the error detection
has been tweaked to only update the error counter for actual errors caused
by the remote proxy itself, rather than third-party webservers. Based on that,
the logic to back off from misbehaving servers has been tweaked, so that it
starts rerouting to other servers earlier. To control this behavior better,
the rules are not split into 'softrules' and 'hardrules'. Connections matching
a softrule can be rerouted to another server when needed; connections matching
a hardrule cannot and will fail if the server is unavailable. Lastly, a config
item 'loglevel' has been added to control the amount of stdout output from the
application. The options are, in increasing order of verbosity, 'error', 'info'
and 'debug', with 'info' set by default.